requirements.txt
aws-cdk-lib==2.66.1
constructs>=10.0.0,<11.0.0
jupyter_server==2.4.0
python-dotenv==1.0.0
boto3==1.26.87
